Refactor setup function in multi-holder, single issuer scenarios
