+++ scripts/k6/scripts/collection_proof.sh
@@ -5,15 +5,16 @@ set -euo pipefail
 source "$(dirname "${BASH_SOURCE[0]}")/common.sh"
 
 config() {
-  export VUS=10
+  export VUS=5
   export ITERATIONS=10
   export ISSUER_PREFIX="k6_issuer_proof"
   export HOLDER_PREFIX="k6_holder_proof"
 }
 
 init() {
+    xk6 run ./scenarios/bootstrap-issuer.js -e ITERATIONS=1 -e VUS=1
     run_test ./scenarios/create-holders.js
-    run_test ./scenarios/create-invitation.js
+    run_test ./scenarios/create-invitations.js
     run_test ./scenarios/create-credentials.js
 }
 
@@ -22,10 +23,10 @@ scenario() {
 }
 
 cleanup() {
-    echo "Cleaning up..."
-    local iterations=$((INITIAL_ITERATIONS * INTIAL_VUS))
+    log "Cleaning up..."
+    local iterations=$((ITERATIONS * VUS))
     local vus=1
-    xk6 run ./scenarios/delete-holders.js -e ITERATIONS=${iterations} -e VUS=${vus}
+    xk6 run ./scenarios/delete-holders.js
     xk6 run ./scenarios/delete-issuers.js -e ITERATIONS=1 -e VUS=1
 }
